chore: moved useMultiMediaStreams hook from libs to publisher app
