+++ apps/publisher/src/hooks/use-can-capture-stream/index.ts
