--- libs/use-multi-media-streams/project.json
@@ -1,16 +1,0 @@
-{
-  "name": "use-multi-media-streams",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/use-multi-media-streams/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/use-multi-media-streams/**/*.{ts,js}"]
-      }
-    }
-  }
-}
